feat: implement character behavior on hit by monster
When character hits with a monster, character is knock-back with speed of that monster and have some invincible time monster couldn't touch him.
Revise code so that INVINSIBLE status of CharacterClass is controlled by independent variable 'm_TimeInvincibleEnd',
rather than considering it as state, in order to freely behave even when character is invincible.